+++ poseidon/src/hash.rs
@@ -1,15 +1,16 @@
 use ark_bls12_381::Fr;
 use paste::paste;
 
-use crate::{parameters::*, DOMAIN_SEPARATOR};
+use crate::{domain_separator, parameters::*};
 
 macro_rules! n_to_one {
     ($n: literal, $n_as_word: literal) => {
         paste! {
+            /// Compute Poseidon hash for `input`.
             pub fn [<$n_as_word _to_one_hash>] (input: [Fr; $n]) -> Fr {
-                let parameters = [<RATE_ $n _PARAMETERS>].clone();
+                let parameters = [<rate_ $n>]::<Fr>();
                 let mut state = poseidon_permutation::Instance::new(&parameters);
-                state.n_to_1_fixed_hash([ark_std::vec![*DOMAIN_SEPARATOR], input.to_vec()].concat())
+                state.n_to_1_fixed_hash([ark_ff::vec![domain_separator()], input.to_vec()].concat())
             }
         }
     };
